fix: simplify verify script to check config loading only
